Update standard returns reminder
